change code style again
