+++ src/main/java/ru/mail/polis/ads/part1/bogdanmendli/FirstTask.java
@@ -1,4 +1,4 @@
-package ru.mail.polis.ads.part1.bogdanMendli;
+package ru.mail.polis.ads.part1.bogdanmendli;
 
 import java.io.PrintWriter;
 import java.util.Scanner;
@@ -8,15 +8,14 @@ public final class FirstTask {
     private FirstTask() {
     }
 
-    private static void solve(final Scanner in, final PrintWriter out) {
+    private static void solve() {
+        final Scanner in = new Scanner(System.in);
+        final PrintWriter out = new PrintWriter(System.out);
         final int result = in.nextInt();
         out.println(result / 10 + " " + result % 10);
     }
 
     public static void main(final String[] args) {
-        try (PrintWriter out = new PrintWriter(System.out);
-             Scanner in = new Scanner(System.in)) {
-            solve(in, out);
-        }
+        solve();
     }
 }
