+++ src/main/java/ru/mail/polis/ads/part1/bogdanmendli/SecondTask.java
@@ -1,21 +1,27 @@
-package ru.mail.polis.ads.part1.bogdanMendli;
+package ru.mail.polis.ads.part1.bogdanmendli;
 
 import java.io.PrintWriter;
 import java.util.Scanner;
 
 public final class SecondTask {
 
+    private static StringBuilder minSequence = new StringBuilder();
+    private static String[][] resultSequences;
+    private static String sequence;
+
     private SecondTask() {
     }
 
-    private static void solve(final Scanner in, final PrintWriter out) {
-        final String sequence = in.nextLine();
+    private static void solve() {
+        final Scanner in = new Scanner(System.in);
+        final PrintWriter out = new PrintWriter(System.out);
+        sequence = in.nextLine();
         if ("".equals(sequence)) {
             out.println();
             return;
         }
         final int length = sequence.length();
-        String[][] resultSequences = new String[length][length];
+        resultSequences = new String[length][length];
         for (int i = 1; i <= length; i++) {
             for (int left = 0; left + i - 1 < length; left++) {
                 final int right = left + i - 1;
@@ -34,24 +40,24 @@ public final class SecondTask {
                         resultSequences[left][right] = resultSequences[left][left] + resultSequences[right][right];
                     }
                 } else {
-                    String minSequence = "0";
-                    for (int k = 0; k < 20; k++) {
-                        minSequence.concat("0000000000");
-                    }
+                    minSequence.delete(0, minSequence.length());
+                    minSequence.append("0".repeat(101));
                     for (int right1 = left; right1 < right; right1++) {
                         final int left2 = right1 + 1;
                         final String temp = resultSequences[left][right1] + resultSequences[left2][right];
                         if (temp.length() < minSequence.length()) {
-                            minSequence = temp;
+                            minSequence.delete(0, minSequence.length());
+                            minSequence.append(temp);
                         }
                     }
-                    if ((sequence.charAt(left) == '(' && sequence.charAt(right) == ')')
-                        || (sequence.charAt(left) == '[' && sequence.charAt(right) == ']')) {
+                    if (sequence.charAt(left) == '(' && sequence.charAt(right) == ')'
+                        || sequence.charAt(left) == '[' && sequence.charAt(right) == ']') {
                         final String temp = sequence.charAt(left)
                             + resultSequences[left + 1][right - 1]
                             + sequence.charAt(right);
                         if (temp.length() < minSequence.length()) {
-                            minSequence = temp;
+                            minSequence.delete(0, minSequence.length());
+                            minSequence.append(temp);
                         }
                     }
                     resultSequences[left][right] = minSequence.toString();
@@ -62,9 +68,6 @@ public final class SecondTask {
     }
 
     public static void main(final String[] arg) {
-        final Scanner in = new Scanner(System.in);
-        try (PrintWriter out = new PrintWriter(System.out)) {
-            solve(in, out);
-        }
+        solve();
     }
 }
